Added Change from Jukka
it seems that the problem is here:
https://github.com/learning-layers/openldapaccount-Dockerfiles/blob/master/docker-compose.yml

’links:’ is defined twice, once as
-openldap
-oidc

and second time:
-openldap

It seems that docker-compose versions deal with this in different ways. …And when layersbox -script reads this, it ends up writing this kind of definition:

openldapaccount:
... links: [openldap]...
 
if this is replaced with:
  links: [openldap, oidc]

Then the registration works.
So if you can remove the second ’links’ definition in openldapaccount’s docker-compose.yml , layersbox script should be able to write it the correct way.